move navi intro
