+++ app/component/itinerary/ItineraryPage.js
@@ -74,7 +74,7 @@ import {
 } from './ItineraryPageUtils';
 import ItineraryTabs from './ItineraryTabs';
 import NaviContainer from './Navigator/NaviContainer';
-import NavigatorIntroModal from './NavigatorIntro/NavigatorIntroModal';
+import NavigatorIntroModal from './Navigator/NavigatorIntro/NavigatorIntroModal';
 import planConnection from './PlanConnection';
 
 const MAX_QUERY_COUNT = 4; // number of attempts to collect enough itineraries
