+++ app/component/itinerary/Navigator/NavigatorIntro/NavigatorIntroFeature.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import { FormattedMessage } from 'react-intl';
-import Icon from '../../Icon';
+import Icon from '../../../Icon';
 
 const NavigatorIntroFeature = ({
   icon,
