+++ app/component/itinerary/Navigator/NavigatorIntro/navigator-intro.scss
