+++ sass/_main.scss
@@ -64,7 +64,7 @@ $body-font-weight: $font-weight-medium;
 @import '../app/component/itinerary/mobile-ticket-purchase-information';
 @import 'base/button';
 @import '../app/component/itinerary/Navigator/navigator';
-@import '../app/component/itinerary/NavigatorIntro/navigator-intro';
+@import '../app/component/itinerary/Navigator/NavigatorIntro/navigator-intro';
 
 /* Modal */
 @import '~foundation-apps/scss/helpers/breakpoints';
